+++ mne/fiff/tests/test_raw.py
@@ -68,6 +68,32 @@ def test_io_raw():
         fname = op.join(op.dirname(__file__), 'data', 'test_raw.fif')
 
 
+def test_io_complex():
+    """ Test IO with complex data types """
+    dtypes = [np.complex64, np.complex128]
+
+    raw = Raw(fif_fname, preload=True)
+    picks = np.arange(5)
+    start, stop = raw.time_to_index(0, 5)
+
+    data_orig, _ = raw[picks, start:stop]
+
+    for dtype in dtypes:
+        imag_rand = np.array(1j * np.random.randn(data_orig.shape[0],
+                            data_orig.shape[1]), dtype)
+
+        raw_cp = deepcopy(raw)
+        raw_cp._data = np.array(raw_cp._data, dtype)
+        raw_cp._data[picks, start:stop] += imag_rand
+        raw_cp.save('raw.fif', picks, tmin=0, tmax=5)
+
+        raw2 = Raw('raw.fif')
+        raw2_data, _ = raw2[picks, :]
+        n_samp = raw2_data.shape[1]
+        assert_array_almost_equal(raw2_data[:, :n_samp],
+                                  raw_cp._data[picks, :n_samp])
+
+
 def test_getitem():
     """Test getitem/indexing of Raw
     """
@@ -122,13 +148,13 @@ def test_filter():
     picks = picks_meg[:4]
 
     raw_lp = deepcopy(raw)
-    raw_lp.low_pass_filter(picks, 4.0)
+    raw_lp.low_pass_filter(picks, 4.0, verbose=0)
 
     raw_hp = deepcopy(raw)
-    raw_hp.high_pass_filter(picks, 8.0)
+    raw_hp.high_pass_filter(picks, 8.0, verbose=0)
 
     raw_bp = deepcopy(raw)
-    raw_bp.band_pass_filter(picks, 4.0, 8.0)
+    raw_bp.band_pass_filter(picks, 4.0, 8.0, verbose=0)
 
     data, _ = raw[picks, :]
 
@@ -143,3 +169,13 @@ def test_filter():
     bp_data, _ = raw_bp[picks_meg[4:], :]
 
     assert_array_equal(data, bp_data)
+
+def test_analytic():
+    """ Test computation of analytic signal """
+    raw = Raw(fif_fname, preload=True)
+    picks_meg = pick_types(raw.info, meg=True)
+    picks = picks_meg[:4]
+
+    raw.analytic_signal(picks, verbose=0)
+
+    #XXX what to test?